requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
mysqlclient==2.1.1
protobuf==3.19.4
pyaes==1.6.1
Pyrogram==2.0.27
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.0
SQLAlchemy==1.4.39
TgCrypto==1.2.3
urllib3==1.26.9
